fix: 移动测试文件到正确位置 (#2751)
