--- tests/modules/analytics/__init__.py
@@ -1,1 +1,0 @@
-# Analytics module tests
